+++ CHANGELOG.md
@@ -7,10 +7,12 @@
 #### New
 
 - Temporary models for buckets
+- Ponder scenes for cobblegen mechanics
 
 #### Fixes
 
 - Removed edge cases where some cobblegen interactions would generate blocks without mana
+- Adds missing icons for creatania ponder categories
 
 ### 1.18.2-0.2.0-alpha